+++ bee-ledger/src/workers/consensus/state.rs
@@ -11,8 +11,8 @@ use crate::{
 
 use bee_message::{
     address::Address,
-    constants::IOTA_SUPPLY,
-    output::{self, dust_outputs_max},
+    constant::IOTA_SUPPLY,
+    output::{self},
 };
 use bee_storage::access::AsIterator;
 
@@ -25,9 +25,12 @@ fn validate_ledger_unspent_state<B: StorageBackend>(storage: &B, treasury: u64) 
         let output = storage::fetch_output(storage, &*output_id)?.ok_or(Error::MissingUnspentOutput(output_id))?;
 
         let amount = match output.inner() {
-            output::Output::SignatureLockedSingle(output) => output.amount(),
-            output::Output::SignatureLockedDustAllowance(output) => output.amount(),
+            output::Output::Simple(output) => output.amount(),
             output::Output::Treasury(_) => return Err(Error::UnsupportedOutputKind(output.kind())),
+            output::Output::Extended(output) => output.amount(),
+            output::Output::Alias(output) => output.amount(),
+            output::Output::Foundry(output) => output.amount(),
+            output::Output::Nft(output) => output.amount(),
         };
 
         supply = supply
@@ -51,10 +54,8 @@ fn validate_ledger_balance_state<B: StorageBackend>(storage: &B, treasury: u64) 
     let mut supply: u64 = 0;
 
     for result in iterator {
-        let (address, balance) = result.map_err(|e| Error::Storage(Box::new(e)))?;
-        if balance.dust_outputs() > dust_outputs_max(balance.dust_allowance()) {
-            return Err(Error::InvalidLedgerDustState(address, balance));
-        }
+        let (_, balance) = result.map_err(|e| Error::Storage(Box::new(e)))?;
+        // TODO check dust ?
         supply = supply
             .checked_add(balance.amount())
             .ok_or_else(|| Error::LedgerStateOverflow(supply as u128 + balance.amount() as u128))?;