+++ bee-test/src/rand/metadata.rs
@@ -12,8 +12,10 @@ use bee_tangle::{
 };
 
 /// Generates a random conflict reason.
+/// It leaves out [`ConflictReason::SemanticValidationFailed`] as it is just a placeholder defined by the protocol but
+/// is not actually being used within the bee framework.
 pub fn rand_conflict_reason() -> ConflictReason {
-    ((rand_number::<u64>() % 7) as u8).try_into().unwrap()
+    ((rand_number::<u64>() % 6) as u8).try_into().unwrap()
 }
 
 /// Generates a random message metadata.