+++ bee-tangle/CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
+## 0.3.0 - 2022-XX-XX
+
+### Removed
+
+- `ConflictReason::InvalidDustAllowance`;
+
 ## 0.2.0 - 2021-12-07
 
 ### Changed